Fix heap-buffer-overflow and memory leak
First, in every function, we should consider the input parameter. We
should answer the quetion that whether or not the pointer is point to
NULL? In that case, we should not visit it content which can cause
**Buffer heap overflow**.In any case, we should not asumme the input is
good. We should always handle the bad input.

Second, the sizeof(<pointer>) return the size of a pointer, not the
memory size it point to. This the differences between the sizeof and
stelen.

Last, for every molloc we should consider free it. In a function that
has exceptional case, we do not link it malloced memory to the linked
list, in that case we should free the malloced memory.